pyproject.toml
[project]
name = "ign8"
version = "4.7.60"
description = "Knowit automation lifecycle management"
readme = "README.md"
requires-python = ">=3.9"
keywords = [ "ign8", "awx", "ansible", "automation",]
classifiers = [ "Development Status :: 3 - Alpha", "Intended Audience :: Developers", "Topic :: Software Development :: Libraries :: Python Modules", "License :: OSI Approved :: MIT License", "Programming Language :: Python :: 3",]
dependencies = [ "redis", "flask", "pynetbox", "wheel", "hvac", "xmltodict", "netbox", "urllib3", "cryptography", "PyYAML", "toml", "django", "djangorestframework", "gunicorn", "ansible", "ansible-core", "paramiko",]
[[project.authors]]
name = "Jakob Holst"
email = "jho@miracle.dk"

[[project.maintainers]]
name = "Jakob Holst"
email = "jho@miracle.dk"

[build-system]
requires = [ "setuptools>=43.0.0", "wheel",]
build-backend = "setuptools.build_meta"

[project.license]
file = "LICENSE.txt"

[project.optional-dependencies]
dev = [ "check-manifest",]
test = [ "coverage",]

[project.urls]
Homepage = "https://github.com/JakobHolstDK/ign8"
"Bug Reports" = "https://github.com/JakobHolstDK/ign8/issues"
Source = "https://github.com/JakobHolstDK/ign8"

[project.scripts]
ign8 = "ign8:main"
ign8_airflow = "ign8.airflow:main"
ign8_bump = "ign8.bump:main"
ign8_dns = "ign8.dns:main"
ign8_file = "ign8.file:main"
ign8_gitea = "ign8.gitea:main"
ign8_iad = "ign8.iad:main"
ign8_inabox = "ign8.inabox:main"
ign8_jenkins = "ign8.jenkins:main"
ign8_netbox = "ign8.netbox:main"
ign8_pitv = "ign8.pitv:main"
ign8_selinux = "ign8.selinux:main"
ign8_semaphore = "ign8.semaphore:main"
ign8_terraform = "ign8.terraform:main"
ign8_traefik = "ign8.traefik:main"
ign8_ui = "ign8.ui:main"
ign8_vault = "ign8.vault:main"
ign8_zabbix = "ign8.zabbix:main"

[tool.poetry]
name = "ign8"
version = "4.7.60"
description = "Ignite it all."
readme = "README.md"
authors = [ "Jakob Holst <jakob.holst@knowit.dk>",]
license = "MIT"
homepage = "https://ign8.openknowit.com"
repository = "https://github.com/miracle-as/openknowit_ign8.git"

[tool.poetry.scripts]
ign8 = "ign8:main"
ign8_airflow = "ign8.airflow:main"
ign8_bump = "ign8.bump:main"
ign8_dns = "ign8.dns:main"
ign8_file = "ign8.file:main"
ign8_gitea = "ign8.gitea:main"
ign8_iad = "ign8.iad:main"
ign8_inabox = "ign8.inabox:main"
ign8_jenkins = "ign8.jenkins:main"
ign8_netbox = "ign8.netbox:main"
ign8_pitv = "ign8.pitv:main"
ign8_selinux = "ign8.selinux:main"
ign8_semaphore = "ign8.semaphore:main"
ign8_terraform = "ign8.terraform:main"
ign8_traefik = "ign8.traefik:main"
ign8_ui = "ign8.ui:main"
ign8_vault = "ign8.vault:main"
ign8_zabbix = "ign8.zabbix:main"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25"
pytest = "^6.2"
mypy = "^0.910"
redis = "^4.5.3"
pynetbox = "^6.6.2"
hvac = "^1.1.0"
netbox = "^0.0.2"
python-jenkins = "^1.7.0"
xmltodict = "^0.13.0"
PyYAML = "^6.0.1"
cryptography = "^41.0.2"
paramiko = "^3.3.1"
toml = "^0.10.2"
Django = "^4.2.8"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
ansible = "^8.7.0"
ansible-core = "^2.15.8"

[tool.setuptools.package-data]
ign8 = [ "*.dat",]
